refactor(tests): move Switchio model tests under `enrollment_switchio`
